refactor: move supporting functions to `utils`
