+++ src/utils/validateFieldType.test.ts
@@ -2,24 +2,24 @@ import { describe, expect, it } from "vitest";
 
 import type { FieldSpec } from "../types.js";
 
-import { validateType } from "./validateType.js";
+import { validateFieldType } from "./validateFieldType.ts";
 
-describe("validateType", () => {
+describe("validateFieldType", () => {
 	it("should return an empty array if no type or types are defined", () => {
 		const field = {} satisfies FieldSpec;
-		const result = validateType("testField", field, "testValue");
+		const result = validateFieldType("testField", field, "testValue");
 		expect(result).toEqual([]);
 	});
 
 	it("should validate a single type correctly", () => {
 		const field = { type: "string" } satisfies FieldSpec;
-		const result = validateType("testField", field, "testValue");
+		const result = validateFieldType("testField", field, "testValue");
 		expect(result).toEqual([]);
 	});
 
 	it("should return an error if the value does not match the single type", () => {
 		const field = { type: "string" } satisfies FieldSpec;
-		const result = validateType("testField", field, true);
+		const result = validateFieldType("testField", field, true);
 		expect(result).toEqual([
 			"Type for field testField was expected to be string, not boolean",
 		]);
@@ -27,15 +27,15 @@ describe("validateType", () => {
 
 	it("should validate multiple types correctly", () => {
 		const field = { types: ["string", "boolean"] } satisfies FieldSpec;
-		const result1 = validateType("testField", field, "testValue");
-		const result2 = validateType("testField", field, true);
+		const result1 = validateFieldType("testField", field, "testValue");
+		const result2 = validateFieldType("testField", field, true);
 		expect(result1).toEqual([]);
 		expect(result2).toEqual([]);
 	});
 
 	it("should return an error if the value does not match any of the multiple types", () => {
 		const field = { types: ["string", "boolean"] } satisfies FieldSpec;
-		const result = validateType("testField", field, {});
+		const result = validateFieldType("testField", field, {});
 		expect(result).toEqual([
 			"Type for field testField was expected to be string or boolean, not object",
 		]);
@@ -43,13 +43,13 @@ describe("validateType", () => {
 
 	it("should validate array type correctly", () => {
 		const field = { type: "array" } satisfies FieldSpec;
-		const result = validateType("testField", field, []);
+		const result = validateFieldType("testField", field, []);
 		expect(result).toEqual([]);
 	});
 
 	it("should return an error if the value does not match the array type", () => {
 		const field = { type: "array" } satisfies FieldSpec;
-		const result = validateType("testField", field, "notAnArray");
+		const result = validateFieldType("testField", field, "notAnArray");
 		expect(result).toEqual([
 			"Type for field testField was expected to be array, not string",
 		]);
@@ -57,13 +57,13 @@ describe("validateType", () => {
 
 	it("should validate boolean type correctly", () => {
 		const field = { type: "boolean" } satisfies FieldSpec;
-		const result = validateType("testField", field, true);
+		const result = validateFieldType("testField", field, true);
 		expect(result).toEqual([]);
 	});
 
 	it("should return an error if the value does not match the boolean type", () => {
 		const field = { type: "boolean" } satisfies FieldSpec;
-		const result = validateType("testField", field, "notABoolean");
+		const result = validateFieldType("testField", field, "notABoolean");
 		expect(result).toEqual([
 			"Type for field testField was expected to be boolean, not string",
 		]);
@@ -71,13 +71,13 @@ describe("validateType", () => {
 
 	it("should validate object type correctly", () => {
 		const field = { type: "object" } satisfies FieldSpec;
-		const result = validateType("testField", field, {});
+		const result = validateFieldType("testField", field, {});
 		expect(result).toEqual([]);
 	});
 
 	it("should return an error if the value does not match the object type", () => {
 		const field = { type: "object" } satisfies FieldSpec;
-		const result = validateType("testField", field, "notAnObject");
+		const result = validateFieldType("testField", field, "notAnObject");
 		expect(result).toEqual([
 			"Type for field testField was expected to be object, not string",
 		]);
