+++ src/utils/validatePeople.ts
