+++ src/validators/index.ts
@@ -1,7 +1,4 @@
 export { validateAuthor } from "./validateAuthor.js";
 export { validateBin } from "./validateBin.js";
 export { validateDependencies } from "./validateDependencies.js";
-export { validatePeople } from "./validatePeople.js";
-export { validateType } from "./validateType.js";
 export { validateUrlOrMailto } from "./validateUrlOrMailto.js";
-export { validateUrlTypes } from "./validateUrlTypes.js";
