+++ src/validators/validateAuthor.test.ts
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
+import { isPerson, validatePeople } from "../utils/index.js";
 import { validateAuthor } from "./validateAuthor.js";
-import { isPerson, validatePeople } from "./validatePeople.js";
 
-vi.mock("./validatePeople", () => ({
+vi.mock("../utils/validatePeople", () => ({
 	isPerson: vi.fn(),
 	validatePeople: vi.fn(),
 }));
